Updated + javadoc-ed Problem 2 raw solution

+++ src/main/java/ro/robertmunteanu/eulerproject/util/AssertUtils.java
@@ -12,7 +12,7 @@ public class AssertUtils {
     // Class methods
     // --------------------------------------------------------------------------------------------
 
-    public static void assertEquals(int actual, int expected) {
+    public static void assertEquals(long actual, long expected) {
 
         if (actual != expected)
             throw new IllegalArgumentException(actual + " != " + expected);
